requirements.txt
selenium
python-igraph